Stop truncating node ids to 32 bits
node_uid is a uint64_t counter on ASTInternal, but both GetNodeNum()
accessors returned unsigned, so every reader saw it modulo 2^32.  The
counter is a thread-local incremented by two per node and is never
reset - not per STPMgr, not per query - so it accumulates over a
thread's whole lifetime.  A process that creates 2^31 nodes wraps it.
That is out of reach for the command line tool, which does one solve
per process, but not for long-lived library use: repeated vc_query on
one VC, the language bindings, or a symbolic execution engine driving
STP for hours.

The accessor is not where the damage would be done.  ASTInterior sets
a NOT node's uid to children[0].GetNodeNum() + 1, so the truncated
value is stored, not merely observed.  Past the wrap a NOT node is
minted with a small uid colliding with a live older node's, breaking
the "a node's uid exceeds its children's" invariant that the
NOT(NOT(x)) ban in HashingNodeFactory, the exprless ordering and an
assert in ASTmisc all rely on.  Widening the accessor fixes that site
by construction.

Containers keyed on an ASTNode are unaffected: Hash() already reads
the full 64 bits and ASTNodeEqual compares pointers.  Only containers
keyed on the integer were exposed, and a collision there is a silent
wrong answer rather than a crash - Flatten's fromTo would hand back
the wrong replacement node, and Dependencies would merge two nodes'
parent sets.

Most of those containers already declared uint64_t keys and were
simply being fed a truncated value.  The ones that did not are
widened here: MergeSame's index map, VariablesInExpression's symbol
graph, DifficultyScore's memo, the two graph printers' visited sets,
and the read counter in ASTmisc.  Cpp_interface::Entry::node_number
becomes a uint64_t with an explicit NO_NODE sentinel, replacing an
int -1 that was compared through an unsigned cast.

getExprID is widened to uint64_t as well, which changes the C API.  It
is the one caller outside the library, its result is an opaque
identity, and returning a truncated one defeats the point of the fix.
The ctypes binding is updated to match.

A -Wconversion sweep over lib/ and tools/ reports no remaining
narrowing of a node id.

+++ lib/Printer/GDLPrinter.cpp
@@ -41,7 +41,7 @@ using namespace stp;
 void outputBitVec(const ASTNode n, ostream& os);
 
 void GDL_Print1(ostream& os, const ASTNode& n,
-                std::unordered_set<int>* alreadyOutput,
+                std::unordered_set<uint64_t>* alreadyOutput,
                 string (*annotate)(const ASTNode&))
 {
   // check if this node has already been printed. If so return.
@@ -134,7 +134,7 @@ ostream& GDL_Print(ostream& os, const ASTNode n,
   os << "display_edge_labels: yes" << endl;
 
   // create hashmap to hold integers (node numbers).
-  std::unordered_set<int> alreadyOutput;
+  std::unordered_set<uint64_t> alreadyOutput;
 
   GDL_Print1(os, n, &alreadyOutput, annotate);
   ;
